Step 17 Added test and clean up

+++ data/src/main/java/com/coroutinelab/data/dto/emaillist/EmailListItemDto.kt
@@ -22,5 +22,5 @@ data class EmailListItemDto(
     @SerialName("threadId")
     val threadId: String?,
     @SerialName("timestamp")
-    val timestamp: Long?
+    val timestamp: Int?
 )
